+++ vtm/src/org/oscim/utils/Parameters.java
@@ -22,6 +22,11 @@ public final class Parameters {
     public static boolean ANIMATOR2 = false;
 
     /**
+     * Allow custom scale factor for short vertices instead of the calculated one.
+     */
+    public static boolean CUSTOM_COORD_SCALE = false;
+
+    /**
      * Allow custom tile size instead of the calculated one.
      */
     public static boolean CUSTOM_TILE_SIZE = false;